Use TextureView for effects test fragment
It's hard to get a meaningful test result with XCID.  I noticed this line was accidentally commented out. I wonder it would make a difference. The SIGSEGV could be from writing to a Surface that no longer exists, which could happen with SurfaceView because we don't control the Surface with SurfaceView.

Bug: 285396965
Test: manual test and ./gradlew bOS
Change-Id: I2c46888b3019bd82c85419767828bdaf65c7f7f9