Move most of prefetchTest to the main thread
This test has been flaking in a way that suggests there is a concurrency
problem.  This change moves everything that's not a blocking call to the
main thread so that all the relevant reads and writes will occur on one
thread, rather than being split between the main and test threads.

Note: I have not been able to reproduce this failure locally, so this
fix is thus somewhat speculative.

Fixes: 285199733
Test: Ran the test locally and it passed
Change-Id: I0888614a9a6e7455207f2bcd65910c6f28576b7e