Fix flaky PreviewViewFragmentTest
Use the DeferrableSurface from the SurfaceRequest directly fixes the problem. This is also consistent with VideoCapture.

However we still don't understand why it happend in the first place. Closing the parent DeferrableSurface(SurfaceEdge.SettableSurface) should close the SurfaceRequest's DeferrableSurface too. The difference might be that there is a delay waiting for the camera to release the Surface. We will fix the failing tests first before investigating further.

Bug: 286817690
Test: manual test and ./gradlew bOS
Change-Id: Ic1252294a481df65f2fe1f7e53ace646192e69bf