[GH] Remove usage of MainDispatcherRule in PagingDataDifferTest
Opted to just pass in an `EmptyCoroutineContext` to `PagingDataDiffer` instead of inlining `MainDispathcherRule`, as it's clearer to follow. The removal of `loadDispatcher` was necessary for this to pass. I didn't trace why this had to be removed, but imo it makes the tests easier to read.

Test: ./gradlew test connectedCheck
Bug: 270612487

This is an imported pull request from https://github.com/androidx/androidx/pull/562.

Resolves #562
Github-Pr-Head-Sha: 1f5051a8905d7b5048cfe6c809f9158f5965f33d
GitOrigin-RevId: 469c54392eb6c6a02dd28af1f2921d7bc80b023b
Change-Id: Ic697668277be5ad8d7f99c244a5977127f4f96ab