Improve snapping animations in Pager
Updated how snapping animations work to make them more consistent. We updated the SnapLayoutInfoProvider for Pager to make the animation specs used more deterministic. We also optimized the animations for 1 pager Pagers, the most common use case. In those cases we will always use the same animation spec (usually a spring) to run the settling animation. For multi page snapping, the specs used will vary but we also updated the fling behavior documentation to make it clear which specs should be used to control the settling animations in that case as well.

Relnote: N/A
Test: N/A
Fixes: 272110321
Change-Id: I553c8c67bd1c5323f758ce033b9cc4b622b59661