Pass apiSince JSON to Dackka
This creates a new configuration that can be used in the build.gradle for docs-public or docs-tip-of-tree called "apiSinceDocs". Using it for a dependency instead of "docs" means gradle will try to resolve the version metadata file for that dependency. This will allow us to control which libraries we release docs for with apiSince. Once the feature is fully rolled out we'll want the "docs" configuration to do what "apiSinceDocs" does.

Test: cl/543487132 stages the tip-of-tree docs with apiSince turned on for all non-KMP libraries using aosp/2640774. cl/544053287 stages the public docs with apiSince turned on for navigation (see e.g. https://android.devsite.corp.google.com/reference/androidx/navigation/ActivityNavigator?db=androidx-refdocs) using aosp/2640775
Bug: 287242156

Change-Id: I35a710ccea420394028a9bd65d3d64fbced032b1