Add unit tests for transactional failure cases.
These unit tests cover the following callbacks (in CallControlCallback)
when the client rejects the requests: onSetActive, onSetInactive,
onDisconnect. The tests are being verified for both V2 and backwards
compatibility. Note that the test to reject the onAnswer callback has
already been included as part of a separate CL.

From additional feedback, I've added test coverage for the case when we
reject the disconnect callback via Call.reject. Another note, I came
across test failures when running the entire suite which were happening
as a result of not sending an explicit local disconnect for the call.
This was causing subsequent tests to fail out.

Fixes: 284992597
Test: BasicCallControlCallbacksTest
Change-Id: Iff86c3a5ce987c481a50961fc4da61824508a1fa