Expose watch face IntDef constants
Prior to ag/2470904 IntDefs in the watch face library had publically
visible constants. That patch was correct to restrict the IntDefs,
the problem is the constants shouldn't have been defined in a
companion object of the IntDef.  Unfortunately that ship has sailed
but here we can expose the constants, and update our samples
accordingly. The constants on the IntDef companion objects have
been retained in case there's code that uses them.

Test: Presubmit
Relnote: This ptach exposes WatchFaceType constants in WatchFaceTypes, CanvasType constants in CanvasTypes, TapType contants in TapTypes and ComplicationsSlotBoundsType constants in ComplicationsSlotBoundsType.
Bug: 288750666
Change-Id: I3b85afb17aa66668bb1b012c1eb9849b38affeb9