Always write local.properties files
This makes it so Gradle does not consider it as part of
input to configuration cache. This is safe for us to do
because these values are loaded by Gradle itself.

This is required for Gradle 8.1 adoption as it treats this
file as input to configuration cache if we check for its
existence.

Test: ./gradlew bOS --dry-run
Change-Id: I9bfa8fb17ca5551b2737361b51d4163bb2372168