[Compose] Introduce optimization strategies for MotionLayout
Added InvalidationStrategy class. It allows the user to optimize two
cases of reduced performance in MotionLayout due to invalidation.

The change includes extended documentation explaining why invalidation
happens and full examples on how to mitigate it with
InvalidationStrategy.

Relnote: "You may now provide an InvalidationStrategy to optimize
invalidations in MotionLayout. Which will typically be the source of
reduced performance in MotionLayout"

Bug: 290055201
Test: MotionLayoutTest
Change-Id: Iea5178a12707323c6339baa0bef19c4d9e2864b1