Transition API changes from feedback
Fixes: 287076423

Relnote: "Added setCurrentFraction() to
allow setting the progress of a transitions.

Added TransitionManager.seekTo() to allow
using Scenes for transitions.

Added add/remove OnProgressChanged for
TransitionSeekController so the progress
can be observed from animateToStart/End.

Added physics-based animations to seeking
transitions. It uses a 1-D velocity tracker to
track the progress change with setCurrentFraction()
or setCurrentPlayTimeMillis() and uses it
for the initial velocity of animateToStart/End.
"

Test: new tests, manual testing for visuals
Change-Id: I6d4e743fb1490c6b6d2936b0a5c753c22ab96bf5