[Tooltip] foundation version of tooltips
Adding BasicTooltipBox API into compose.foundation, that is used by both Android and Desktop. It handles the default gestures like long press for touch and stylus as well as  hover for mouse.
Updated Material3 PlainTooltipBox and RichTooltipBox to just be TooltipBox with new PlainTooltip and RichTooltip composables that can be passed into TooltipBox. The new TooltipBox uses
BasicTooltipBox, but also wraps the transitions needed for Material3 spec. Updating relevant tests and demos with these new APIs.

Bug: 287296645
Test: Adding default gesture handling tests for Android BasicTooltipBox. Updating Material3 tooltip unit and screenshot tests to use the new TooltipBox API.
Relnote: Added BasicTooltipBox to compose.foundation and updated PlainTooltipBox and RichTooltipBox to use TooltipBox with new PlainTooltip and RichTooltip composables.
Change-Id: I79e1df0ac02fdccc3399dcf8d24a515d6461fde9