Make progress a parameter for drawing, not a property
A Morph object's progress basically exists to draw the morph at that
progress point (for animations, etc). This change is a simplification
which removes the necessity to first set the progress on a Morph, then
draw it (which uses that progress). Instead, you draw a Morph with
a given progress; it recalculates the path/etc as necessary
given the progress value and then draws it.

Bug 290108274
Test: intergration and unit tests all pass
Relnote: progress is now passed to Morph drawing commands directly

Change-Id: Icdca2f2ced00bd7957fc5dc318298fd083a74586