Add GH CODEOWNERS for PR triage
This allows GH to automatically assign reviewers to PRs, which should
reduce the amount of incoming e-mails by allowing folks on triage to
filter out all PR notifications except those they are explicitly tagged
or assigned to.

CODEOWNERS requires write permissions on the repo, so we had to relax
this permission - but it shouldn't affect anything as we have branch
protections on androidx-main in GitHub.

Test: GH lint
Change-Id: I67f6a502943641afbbf04220e25dfdae7f9fa1ff