Store Compose transition lambdas in destinations
To ensure that all EnterTransition and ExitTransition
lambdas are cleaned up when the NavHost leaves composition,
all information created as part of the Kotlin DSL needs to
be stored within the destinations themselves, rather than
separately.

This means that the transition information associated with
both single destinations and with entire graphs need to be
stored on their respective NavDestination subclassses.

By necessity, this requires a new ComposeNavGraphNavigator that
extends the base NavGraphNavigator. This is currently just
an implementation detail and not exposed externally.

Relnote: "Fixed an issue where the `EnterTransition` and
`ExitTransition` lambdas created as part of the `NavHost` could
potentially remain in memory even after the `NavHost` is removed
from composition."
Test: verified in updated sample

Change-Id: I893d0a0e36b512afe60a5af892cd84995af02015