Upload truth apis as previous stable version
Plan is to use these files to diff against current APIs in Kruth so we can track which parts of the API surface we are still missing to get code complete.

API Council please skip review on these APIs, as Kruth has no plans to
do a public release and we're only using API tracking for source compatibility / feature parity against Truth.

Test: ./gradlew kruth:kruth:checkApi
RelNote: n/a
Change-Id: I026e27d805af9260754d15c9441778f3eef4a2f9