Add a strong skipping mode option
This CL adds an option to the plugin to enable relaxed skipping rules for composables.
It achieves this by removing the assumption that unstable parameters will cause a
composable to be unskippable. When an unstable parameter is encountered when evaluating
if a composable can be skipped, instance equality is used.

When strong skipping mode is enabled, this CL also makes it so lambdas with unstable
captures are now memoized as well.

This configuration will default to off until we evaluate the effects of this change.

Also adds a new NonSkippableComposable annotation to opt a composable out of skipping.

Test: StrongSkippingModeTests
Relnote: Add a compiler option to skip composables with unstable parameters.

Change-Id: I610f9b2703b84bd93cfaa0d97defbee313023830