Update Pager prefetching behavior to match ViewPager2
The number of pages retained when using ViewPager2 was different from Pager in Compose. Therefore, the number of prefetched pages in Pager was updated to match the behavior in Views. Also, the documentation was fixed to make it clear that there will be pages automatically pre-fetched more than specified in beyondBoundsPageCount.

Fixes: 289088847
Test: Fixed previous tests that were affected by the change.
Relnote: Fixed an issue where the prefetching in Pager did not match the behavior in Views.
Change-Id: I93352e6de29fa8bc23b3af7516872d3e2da7ce0e