Eliminate Android dependecies.
Eliminate all android dependencies, creating platforma agnostic APIS.
* All path generating and drawing code is removed, and should be
implemented on each platform.
* All references to PointF android where removed, on the APIs we already
use Float. An internal Point class (inspired on Compose's Offset) is used to
simplify the implementation, but may be removed in the future and
replaced with just Floatsi too.
* To transforms cubics and rounded polygons a new PointTransformer
functional interface is added, that takes a MutablePoint (also a new
interface) and modifies it in-place. This is done mainly to avoid
costly object creation that will be required on Java to return a Point.

Later, this will have wrappers for Android and Compose.

Test: Manual + Unit Tests
Bug: 292289543
RelNote: "Eliminate android dependencies"

Change-Id: Iadc1c32dfc40a3e88e4c8516c59b21e71366e019