Add XMethodType.typeVariables APIs
Also deprecate XMethodType.typeVariableNames since it is replaceable with XMethodType.typeVariables along with `asTypeName().toJavaPoet()`.

KSP does not model type variables as Javac does, it only models declaration (element), and through a KSTypeReference it links to the type var, therefore the implementation of the API uses as bare-bone implementation of XType instead of inheriting from KspType.

Bug: 247851395
Test: ./gradlew room:room-compiler-processing:test
Change-Id: I0a14d9688ac167d3f8582385c9829bfbdcb8a660