Retrieve test semantics on UI thread
Semantic computation triggers an `observeReads` scope that uses SnapshotStateObserver, that does not support multithreading environment.

If measure/draw happens at the same time, it causes a race with state read/writes misattributed to incorrect scopes.

The fix is to ensure that semantic node computation is always running on UI thread.

Fixes: 293259558
Change-Id: Ida4d112eadd7f2dfb25136ecadc5f04949cffd35