Introduce new infrastructure for multi process tests
Current MultiProcessStore-MultiProcess tests rely on a blocking service
connection combined with custom services for each test case.

There are a couple of problems with the current test:
* Combination of java syncronization and coroutines frequently makes these
tests flaky (e.g. my API 32 emulator cannot run them to completion)
* Tests are harder to read due to the built in signaling logic combined with
custom service implementations. It is hard to understand what "signalService"
will do w/o reading the declaration of the service implementation and other
place where it is waiting for a signal.
* We don't clean after tests hence when a service is re-used, it might get
blocked.
* Creating a new test is unnecessarily verbose as it requires 2 service class
implementations (and declarations in the manifest).

This new solution provides an infrastructure where we can run IpcSubjects
in a remote process and execute IpcActions in their scope. These actions are
parcelable classes that use kotlin's Parcelize for convenience.

There is also a MultiProcessRule that puts all of this together and provides
resource management across tests.

As a result, we can use a single service declaration to run multiple scenerios
and can easily implement them through IpcAction's.

This first CL introduces the infrastructure and coverts one test. Followup
CL's will convert the remaining of tests.

Bug: 294875831
Test: TwoWayIpcTest
Change-Id: I34b717a4e08261e3693ab2b70cb793359361dcfb