Close DeferrableSurfaces after ERROR_CAMERA_SERVICE
When CameraService errors happen, it may fail to disconnect
the Surfaces from the Camera.
Configure the above Surface to the Camera will cause a failure
on the CameraCaptureSession.

There are 2 possible ways to solve the issue
(1) Disconnect the Surfaces after the error of the Camera.
It requires configuring CameraCaptureSession with another Surface
like configAndClose() did.
(2) Create another output Surface for the further usage

We apply the (2) way to have lower latency. Close the DeferrableSurface
(after CameraDevice onError()) to ensure the Surface does not
be configured on the Camera.

Bug: 290861504
Test: CameraControllerFragmentTest, CaptureSessionTest
Change-Id: Ie8804bdc3bb56aae54870875e0d522fd60070498