Support predictive back for transitions in Fragments
Providing support for the new AndroidU Predictive back APIs for AndroidX
Transitions in Framgent. This makes it so that using gesture back with
Androidx Transitions in Framgents is now seekable meaning you can see
the previous destination without completely going back to it.

RelNote: "Fragments now provide support for Predictive back when using
Androidx Transitions. This allows you to use the back gesture to seek to
the previous fragment with your custom Androidx Transition before
deciding to either commit or cancel the transaction via the completed
gesture."
Bug: 285175724
Test: sample app

Change-Id: Ib49b455d0a00d38203de7616fb2be05a3677a548