Fix VelocityTracker test flakiness: VelocityTrackingParityTest
The historical events were being transformed twice and that influenced how the reference point was used in the tracker. Removed the historical transformation. Added a new test to check the parity between lists: flinging a similar list in both Views and Compose should get us to a similar item position.

Fixes: 293188485
Bug: 269487059
Test: Previous Tests Should pass and added new tests.
Change-Id: I3a6b8d84e3cede969716d0468c2f70854cb1434a