Make assertIsDisplayed lenient about bounds larger than screen
The last line of checkIsDisplayed only returned true if the element
bounds are fully contained in the screen. However, in the majority of
real-world tests, this was actually equivalent to checking for partial
visibility. The reason is that usually, most element bounds already got
clipped to the screen by a parent element (either
Modifier.clipScrollableContainer, or a Surface such as the one in
rule.setMaterialContent).  This left only some edge cases where
checkIsDisplayed was truly checking for full visibility instead.

These edge cases have been causing problems (see the two bugs below),
and in general full visibility checking is not a good default, because
some elements are larger than the screen but there should be a way of
checking if they're displayed). So this patch switches
`assertIsDisplayed` explicitly to partial visibility.

Also, add a `strict` parameter to preserve the existing
`assertIsNotDisplayed` behavior, in order not to risk breaking any tests
that may have relied on `assertIsNotDisplayed` passing in these edge
cases.

Test: ProgressIndicatorTest.nonMaterialSetContent
Bug: 295306150
Bug: 267874118

Change-Id: Ibf8039d615234add586fa5e037a86df090b73656