Add Arc based AnimationSpec
Introduces animation using arcs of a quarter of an Ellipse.

- ArcAnimationSpec for single shot animations
- `using(ArcMode)` in `keyframes` for key-frame based animations

These interpolators are designed to be used with 2-dimensional values.
Particularly, for positional values such as Offset.

The interpolator supports multiple modes:
- ArcAbove: The arc corresponds to the curve of a quarter of an
  Ellipse where the curve is above the center of the Ellipse.
- ArcBelow: The arc corresponds to the curve of a quarter of an
  Ellipse where the curve is below the center of the Ellipse.
- ArcLinear: At typical linear interpolation. This is only expected
  to be used in keyframes, where the user might want to avoid arcs in a
  segment of the animation.

While it may be used with other N-dimensional values, its use is
discouraged since it may lead to unpredictable animation curves. The
lint check `ArcAnimationSpecTypeDetector` addresses that. It's an
Informational Lint check that verifies that the calling type of
`ArcAnimationSpec` is one of: Offset, IntOffset
or DpOffset.

See ArcOffsetDemo.kt for a simplified use-case using keyframes using
ArcMode.

Relnote: "You may now use `ArcAnimationSpec` and `using(arcMode:
ArcMode)` (in keyframes) to
interpolate positional values such as Offset using arcs of quarter of an
Ellipse."

Bug: 299477780
Test: ArcAnimationTest, ArcKeyframeAnimationTest
Change-Id: I154c403a5685f1146c5157738b4331967a7d77f7