Conditionally set AccessibilityNodeInfo.importantForAccessiblity
Previously, Compose set importantForAccessibility = true on all
AccessibilityNodeInfo it created, as the vast majority of semantics are
used for accessibility.  However, this has proved to cause some subtle
bugs when there are testTag-only nodes.

This patch introduces a new factory method AccessibilityKey and
switches it for most of the properties in SemanticsProperties.kt (all
those that are accessed in AndroidComposeViewAccessibilityDelegateCompat
). The net change is that now importantForAccessibility = false in the
following situations:

* Non-merging nodes with only non-accessibility-speakable properties.
    * Of the built-in ones, the key example is testTag.
    * Custom SemanticsPropertyKeys defined outside the UI package
      are also non-speakable.
* Non-merging nodes that are empty: notably, clearAndSetSemantics {}
  and the root of the SemanticsNode tree.

Bug: 286372107
Test: ACVADCTest.testIs[Not]Important*
Change-Id: If25ce25a84c1db565167ecfa567d7d4d0554eaf5