SwipeToReveal with EdgeSwipeToDismiss
`Modifier.edgeSwipeToDismiss` works with scrollable by intercepting
the scroll and fling events and sending them to SwipeToDismissBox
before the child node consumes it. However, SwipeToReveal depends
upon swipeable modifier which is a draggable modifier which does
not emit scroll/fling events. To enable edge swipe with SwipeToReveal
a `NestedScrollDispatcher` is required to dispatch the drag events
as scroll/fling event.

Relnote: """
Add NestedScrollDispatcher with SwipeToReveal which can be used with
Modifier.edgeSwipeToDismiss.
"""
Bug: 293444286
Bug: 293436983
Test: Manual testing

Change-Id: I95774a7df008db12756215dc8a064001008103c5